requirements.txt
# Core Dependencies
numpy>=1.21.0
matplotlib>=3.5.0
requests>=2.28.0
groq>=0.4.0
python-dotenv>=0.19.0

# Optional
scipy>=1.7.0
pandas>=1.3.0

openenv-core>=0.2.0
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.0.0
gradio>=4.0.0
